+++ pwnlib/ui.py
@@ -61,10 +61,6 @@ def yesno(prompt, default=None):
       `True` if the answer was "yes", `False` if "no"
 
     Examples:
-
-    .. doctest::
-       :skipif: branch_dev
-
         >>> yesno("A number:", 20)
         Traceback (most recent call last):
         ...
@@ -83,10 +79,6 @@ def yesno(prompt, default=None):
          [?] is it good 3 [yes/No] False
 
     Tests:
-
-    .. doctest::
-       :skipif: branch_dev
-
         >>> p = testpwnproc("print(yesno('is it ok??'))")
         >>> b"is it ok" in p.recvuntil(b"??")
         True
@@ -148,20 +140,12 @@ def options(prompt, opts, default = None):
       The users choice in the form of an integer.
 
     Examples:
-
-    .. doctest::
-       :skipif: branch_dev
-
         >>> options("Select a color", ("red", "green", "blue"), "green")
         Traceback (most recent call last):
         ...
         ValueError: options(): default must be a number or None
 
     Tests:
-
-    .. doctest::
-       :skipif: branch_dev
-
         >>> p = testpwnproc("print(options('select a color', ('red', 'green', 'blue')))")
         >>> p.sendline(b"\33[C\33[A\33[A\33[B\33[1;5A\33[1;5B 0310")
         >>> _ = p.recvall()
@@ -272,10 +256,6 @@ def pause(n=None):
     r"""Waits for either user input or a specific number of seconds.
 
     Examples:
-
-    .. doctest::
-       :skipif: branch_dev
-
         >>> with context.local(log_level="INFO"):
         ...     pause(1)
         [x] Waiting
@@ -287,10 +267,6 @@ def pause(n=None):
         ValueError: pause(): n must be a number or None
 
     Tests:
-
-    .. doctest::
-       :skipif: branch_dev
-
         >>> saved_stdin = sys.stdin
         >>> try:
         ...     sys.stdin = io.TextIOWrapper(io.BytesIO(b"\n"))
@@ -335,11 +311,7 @@ def more(text):
     Returns:
       :const:`None`
 
-    Tests:
-
-    .. doctest::
-       :skipif: branch_dev
-       
+    Tests:       
         >>> more("text")
         text
         >>> p = testpwnproc("more('text\\n' * (term.height + 2))")